requirements.txt
django==4.2.5
envvirons==14.1.1
djangorestframework==3.15.2
django-filter==25.1
flake8==7.3.0
black==25.11.0
isort==6.1.0
drf-spectacular==0.29.0
psycopg2-binary==2.9.11